+++ src/Form/FormPanel.php
@@ -43,7 +43,7 @@ class FormPanel extends FormDefault
     {
         $this->getButtons()->setHtmlAttribute('class', 'panel-footer');
 
-        $this->setHtmlAttribute('class', 'panel panel-default '.$this->getPanelClass());
+        $this->setHtmlAttribute('class', 'card card-body panel panel-default '.$this->getPanelClass());
 
         parent::initialize();
     }